+++ Lib/test/test_import/data/double_const.py
